P3: define HTTP and connector ctype byte domains
Finding: CLUSTER-ctype-http-plain-char

HTTP extension parsing, ACLK HTTPS and WebSocket response-header normalization, and exporter response handling passed remotely sourced plain-char bytes directly to isspace(), tolower(), or isprint(). On signed-char targets, high bytes promote to negative int values other than EOF and violate the C/POSIX ctype argument contract.

The defect is current and source-reachable through direct and ACLK-carried HTTP requests, ACLK HTTPS and streaming responses, MQTT/WebSocket handshakes, and simple exporter response callbacks. A concrete crash, table overread, corruption, or misclassification remains potential and implementation-dependent; no production incident is claimed.

Convert the exact source byte to uint8_t at all eight boundaries. This preserves the original byte modulo the unsigned-byte domain and matches existing repository precedent. ASCII parsing, header normalization, WebSocket negotiation, 200/204 response recognition, buffer flushing, retry behavior, debug formatting bounds, state transitions, and all caller return contracts remain unchanged.

The runtime cost is one constant-time integer conversion per ctype call, with no new branch, allocation, copy, I/O, lock, lifetime, ownership, API, ABI, configuration, or concurrency change.

Validation: all five translation units passed their generated project compile commands in default, forced-signed-char, and forced-unsigned-char modes (15/15); the internal-checks exporter path compiled; ASan/UBSan all-byte probes passed in both signedness modes with identical output; diff and same-failure scans passed. The full object target was unavailable because the checkout lacks the ACLK proto submodule. All six required independent reviewers approved the exact diff, and both reviewer process and semantic gates passed.
(cherry picked from commit bbd3b46402a3a7f3ddbdc581e67cfc3abe7e47b9)

+++ src/aclk/mqtt_websockets/ws_client.c
@@ -337,7 +337,7 @@ int ws_client_parse_handshake_resp(ws_client *client)
             rbuf_bump_tail(client->buf_read, strlen(WS_HTTP_NEWLINE));
 
             for (int i = 0; hdr->key[i]; i++)
-                hdr->key[i] = tolower(hdr->key[i]);
+                hdr->key[i] = tolower((uint8_t)hdr->key[i]);
 
             if (ws_client_add_http_header(client, hdr))
                 return WS_CLIENT_PROTOCOL_ERROR;
